смена типа

+++ examples/rtos_uart/drivers/serial.c
@@ -86,7 +86,7 @@ signed portBASE_TYPE isSerialSendMessageEmpty (void)
 /** циклом добавляем структуру в буффер */
 void xSerialAddMessageSend(const uint8_t * const cOutChar, TickType_t xBlockTime )
 {
-    for(uint8_t i = 0; i < sizeof(cOutChar); i++)
+    for(uint32_t i = 0; i < sizeof(cOutChar); i++)
     {
         xQueueSendToBack( xCharsForTx, &cOutChar[i], xBlockTime );
     }
